% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure.R
\name{create_figure}
\alias{create_figure}
\title{Creates a figure}
\usage{
create_figure(n_sim, all_bias_sim_long, output_dir = getwd())
}
\arguments{
\item{n_sim}{Number of simulations}

\item{all_bias_sim_long}{Data frame containing information on bias from simulation}

\item{output_dir}{The user's working directory}
}
\description{
This function creates a figure that compares the bias of the four estimators under the different settings.
}
