% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Given the factors and factor loadings, simulate the data}
\usage{
fit_models(
  model,
  n,
  trt,
  k_total,
  t_total,
  variance,
  num_timepoints = NULL,
  embedding_dim,
  ortho,
  qual
)
}
\arguments{
\item{model}{The data generated by the generateModel function}

\item{n}{number of units}

\item{trt}{treatment indicator vector}

\item{k_total}{number of outcomes}

\item{t_total}{number of pre-treatment periods}

\item{variance}{variance of the noise term}

\item{num_timepoints}{Number of training timepoints to use. Defaults to t_total.}

\item{embedding_dim}{Dimension into which the matrix Q embeds the donors, ie, num synthetic donors}

\item{ortho}{Should the method be run using orthogonal matrix Q}

\item{qual}{aspect of the data to which the embedding be applied - donors or outcomes.}
}
\description{
Given the factors and factor loadings, simulate the data
}
