% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateModel.R
\name{generateModel}
\alias{generateModel}
\title{Generates data according to factor model}
\usage{
generateModel(t_total, k_total, trt, mu, rho, n)
}
\arguments{
\item{t_total}{number of pre-treatment periods}

\item{k_total}{number of outcomes}

\item{trt}{treatment indicator vector}

\item{mu}{Factor loadings}

\item{rho}{Mixing coefficient}

\item{n}{number of units}
}
\description{
For each outcome, data are generated from a latent factor model, using the mixing coefficient rho to determine the importance of the common factors.
}
