% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Calculates the RMSE}
\usage{
rmse(Y0, Y1, weights)
}
\arguments{
\item{Y0}{Matrix of control outcomes}

\item{Y1}{Vector of target outcomes}

\item{weights}{The estimated synthetic control weights}
}
\description{
Calculates the RMSE
}
