% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCMbias.R
\name{SCMbias}
\alias{SCMbias}
\title{Calculate the bias in the estimators}
\usage{
SCMbias(Y0, Y1, weights)
}
\arguments{
\item{Y0}{Matrix of control outcomes}

\item{Y1}{Vector of target outcomes}

\item{weights}{The estimated synthetic control weights}
}
\description{
Calculates the difference between the prediction and the truth.
}
