% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_qp.R
\name{synth_qp}
\alias{synth_qp}
\title{Solve the synth QP directly}
\usage{
synth_qp(X1, X0)
}
\arguments{
\item{X1}{Target vector}

\item{X0}{Matrix of control outcomes}
}
\description{
This function is used to solve for the weights w_sep, w_cat, w_avg.
}
